// models/DatasQueryResponseInfo.ts
// Generated by https://quicktype.io

export interface DatasQueryResponseInfo {
    total: number;
    data:  NightLightImage[];
}

export interface NightLightImage {
    id:              number;
    UploadTime:      string;
    satelliteID:     SatelliteID;
    sensor:          Sensor;
    frameIndex:      string;
    productLevel:    ProductLevel;
    timeType:        TimeType;
    imagingTime:     string;
    produceTime:     string;
    dataBits:        string;
    rspath:          string;
    widthInPixels:   number;
    heightInPixels:  number;
    imageColumnGSD:  number;
    imageRowGSD:     number;
    roll:            number;
    pitch:           number;
    yaw:             number;
    solarAzimuth:    number;
    solarElevation:  number;
    CenterLongitude: number;
    CenterLatitude:  number;
    LTLatitude:      number;
    LTLongitude:     number;
    RTLatitude:      number;
    RTLongitude:     number;
    RBLatitude:      number;
    RBLongitude:     number;
    LBLatitude:      number;
    LBLongitude:     number;
    path:            string;
    code:            string;
    name:            string;
}

export enum ProductLevel {
    L2 = "L2",
}

export enum SatelliteID {
    Lj01 = "LJ01",
}

export enum Sensor {
    CMOS = "CMOS",
}

export enum TimeType {
    UTC = "UTC",
}